working on #4

+++ simple-libs/simple-libs-jdbc/src/main/java/tech/danfe/simplelibs/simplejdbc/core/ResultSetUtils.java
@@ -115,7 +115,7 @@ public class ResultSetUtils {
                     return resultSet.getInt(columnLabel);
                 }
                 if (type == long.class) {
-                    return resultSet.getInt(columnLabel);
+                    return resultSet.getLong(columnLabel);
                 }
                 if (type == float.class) {
                     return resultSet.getFloat(columnLabel);
